requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
langchain
langchain-community
langchain-google-genai
langchain-pinecone
langchain-text-splitters
pinecone-client
pypdf
langgraph
duckduckgo-search
ddgs
rank_bm25
mcp